requirements.txt
Django==5.1.2
djangorestframework==3.15.2
gunicorn==22.0.0
whitenoise==6.7.0
python-dotenv==1.0.1
requests==2.32.3
